[Issue #17] chore: Separate class into different file.
